+++ resources/views/judicial/manage/cms/videoList.blade.php
@@ -6,7 +6,11 @@
     </div>
     <div class="panel-body">
         <div class="container-fluid">
-            <a type="button" data-key='none' data-method="add" onclick="videoMethod($(this))" class="btn btn-primary">新增</a>
+            @if(!isset($is_archived))
+                <a type="button" data-key='none' data-method="add" onclick="videoMethod($(this))" class="btn btn-primary">新增</a>
+            @else
+                <button type="button" class="btn btn-danger" data-node="system-archivedMng" onclick="loadContent($(this))">返回归档列表</button>
+            @endif
         </div>
         <hr/>
         <div class="container-fluid">
@@ -24,11 +28,13 @@
                 @foreach($video_list as $video)
                 <tr>
                     <td>
-                        <a href="javascript: void(0) ;" data-key="{{ $video['key'] }}" data-method="show" onclick="videoMethod($(this))">查看</a>
-                        &nbsp;&nbsp;
-                        <a href="javascript: void(0) ;" data-key="{{ $video['key'] }}" data-method="edit" onclick="videoMethod($(this))">编辑</a>
-                        &nbsp;&nbsp;
-                        <a href="javascript: void(0) ;" data-key="{{ $video['key'] }}" data-method="delete" data-title="{{ $video['video_title'] }}" onclick="videoMethod($(this))">删除</a>
+                        <a href="javascript: void(0) ;" data-key="{{ $video['key'] }}" data-method="show" data-archived_key="{{ $archived_key }}" data-archived="{{ (isset($is_archived)&&$is_archived=='yes') ? 'yes' : 'no' }}" onclick="videoMethod($(this))">查看</a>
+                        @if(!isset($is_archived))
+                            &nbsp;&nbsp;
+                            <a href="javascript: void(0) ;" data-key="{{ $video['key'] }}" data-method="edit" onclick="videoMethod($(this))">编辑</a>
+                            &nbsp;&nbsp;
+                            <a href="javascript: void(0) ;" data-key="{{ $video['key'] }}" data-method="delete" data-title="{{ $video['video_title'] }}" onclick="videoMethod($(this))">删除</a>
+                        @endif
                     </td>
                     <td>{{ $video['video_title'] }}</td>
                     <td>{{ $video['video_link'] }}</td>
